requirements.txt
appdirs==1.4.4
asgiref==3.2.7
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
distlib==0.3.1
Django==3.0.7
django-crispy-forms==1.9.2
django-embed-video==1.3.3
django-mathfilters==1.0.0
filelock==3.0.12
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.7.0
importlib-resources==3.0.0
mysqlclient==1.4.6
numpy==1.19.0
opencv-python==4.3.0.36
Pillow==7.1.2
pipenv==2020.8.13
psycopg2==2.8.5
pytube==9.6.0
pytube3==9.6.4
pytz==2020.1
qrcode==6.1
requests==2.24.0
six==1.15.0
split==0.4
sqlparse==0.3.1
typing-extensions==3.7.4.2
urllib3==1.25.9
virtualenv==20.0.30
virtualenv-clone==0.5.4
whitenoise==5.1.0
zipp==3.1.0
